test: split github syncer lifecycle tests
Move syncer lifecycle, cancellation, and dispatch coverage into an external syncertest package. The original github package keeps behavior coverage while the race job can now time this work independently from the remaining sync tests.
